% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_question.R
\name{create_question}
\alias{create_question}
\title{Create a new question in the YAML format.}
\usage{
create_question(question_name, question_path, author_name)
}
\arguments{
\item{question_name}{A string indicating the name of the question.}

\item{question_path}{A string indicating the path to your question file.}

\item{author_name}{A string indicating the name of the author}
}
\value{
Creates a .yaml file in the test directory containing the question name and author information. The file name is the question_name parameter.
Returns the file path of the newly created .yaml file.
}
\description{
This function allows users to creates a new question based on questions in the Rosalind platform.
}
\examples{
create_question("Question 1 - Favorite nucleotide?", "~/.", "Jeremy Fan")
}
